PropertyGetOptions are now able to trigger an API call instead of a JSON dump check. VALID_OUTPUTS is no longer used. Next I'll implement fall-through code so "JSON_FIRST" actually means something, after which I'll complete PropertyStore's transition to using PropertyGetOptions rather than strings.
